feat(data): Add repositories, mappers, and domain models
Adds the complete data and domain layers, including repository implementations, domain models, and mappers for all core features of the application.

**Domain Layer:**
- Adds domain models for all application entities: `User`, `UserRole`, `Subject`, `Question`, `QuestionType`, `Answer`, `Test`, `TestMode`, `TestConfiguration`, `TestAssignment`, `TestAttempt`, `TestResult`, `UserAnswer`, etc.
- Defines repository interfaces (`AuthRepository`, `UserRepository`, `SubjectRepository`, `QuestionRepository`, `TestRepository`, etc.) for abstracting data operations.

**Data Layer:**
- Implements all repository interfaces from the domain layer (e.g., `UserRepositoryImpl`, `TestRepositoryImpl`, `AuthRepositoryImpl`). These implementations handle CRUD operations, data validation, and interaction with DAOs.
- Adds mappers to convert between Room entities and domain models for every entity.
- Implements `AuthRepositoryImpl` with user login/logout logic, session management, and password verification.

**Utilities:**
- Adds `PasswordHasher` using jbcrypt for secure password hashing and verification.
- Adds `SessionManager` using DataStore for persisting and managing the current user session.

**Other Changes:**
- Updates all Room entity primary keys from `Int` to `Long`.
- Adds `jbcrypt` and `datastore-preferences` dependencies.

+++ app/src/main/java/com/octal/examly/data/local/entities/AnswerEntity.kt
@@ -19,9 +19,9 @@ import androidx.room.PrimaryKey
 )
 data class AnswerEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
 
-    val questionId: Int,
+    val questionId: Long,
 
     val answerText: String,
